# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateProposalBody {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  idea: string;
  budget?: string;
  timeline?: string;
}

export interface ProposalResponse {
  id: number;
  name: string;
  email: string;
  idea: string;
  budget?: string;
  timeline?: string;
  createdAt: string;
}

export interface Favorite {
  id: number;
  toolName: string;
  createdAt: string;
}

export interface AddFavoriteBody {
  /** @minLength 1 */
  toolName: string;
}

export interface CreateToolSubmissionBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  pricing: string;
  audience?: string;
  /** @minLength 1 */
  url: string;
}

export interface ToolSubmissionResponse {
  id: number;
  name: string;
  category: string;
  description: string;
  pricing: string;
  audience?: string;
  url: string;
  status: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}