# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Submit a project proposal
 */




export const CreateProposalBody = zod.object({
  "name": zod.string().min(1),
  "email": zod.string().email(),
  "idea": zod.string().min(1),
  "budget": zod.string().optional(),
  "timeline": zod.string().optional()
})


/**
 * @summary List the signed-in user's favorite tools
 */
export const ListFavoritesResponseItem = zod.object({
  "id": zod.number(),
  "toolName": zod.string(),
  "createdAt": zod.coerce.date()
})
export const ListFavoritesResponse = zod.array(ListFavoritesResponseItem)


/**
 * @summary Add a tool to the signed-in user's favorites
 */



export const AddFavoriteBody = zod.object({
  "toolName": zod.string().min(1)
})


/**
 * @summary Remove a tool from the signed-in user's favorites
 */
export const RemoveFavoriteParams = zod.object({
  "toolName": zod.coerce.string()
})


/**
 * @summary Submit a new tool (authenticated)
 */







export const SubmitToolBody = zod.object({
  "name": zod.string().min(1),
  "category": zod.string().min(1),
  "description": zod.string().min(1),
  "pricing": zod.string().min(1),
  "audience": zod.string().optional(),
  "url": zod.string().min(1)
})