# lib/api-zod/src/generated/types/createProposalBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProposalBody {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  idea: string;
  budget?: string;
  timeline?: string;
}