# lib/api-zod/src/generated/types/createToolSubmissionBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateToolSubmissionBody {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  pricing: string;
  audience?: string;
  /** @minLength 1 */
  url: string;
}