# lib/api-zod/src/generated/types/proposalResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProposalResponse {
  id: number;
  name: string;
  email: string;
  idea: string;
  budget?: string;
  timeline?: string;
  createdAt: Date;
}