# lib/api-zod/src/generated/types/toolSubmissionResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ToolSubmissionResponse {
  id: number;
  name: string;
  category: string;
  description: string;
  pricing: string;
  audience?: string;
  url: string;
  status: string;
  createdAt: Date;
}